Fix Old RC removal logic and add test for the function
Old RC releases should be removed when a new RC is built. This helps
us avoid the situation we had with 3.1.4. The step to remove the old RC wasn't working before
because we were checking old releases against Airflow 2 instead of 3
using if name.startswith("2"). The fix here is to use pattern to select
all old RCs whether they are version 2 or 3.