Update API call to use JWT for authentication
I followed this guide, but example to test API is no longer working as "v1" API is deprecated and system requires to use "v2" instead. More importantly, I was unable to authenticate using username and password. It seems that JWT token is the only valid authentication asset. I prepared therefore alternative example which is correcting both issues. Hope it is not too complicated. Cheers, Tomas.