Improve FastApi exclusion for Cadwyn incompatibility and link Cadwyn … (#59859)
* Improve FastApi exclusion for Cadwyn incompatibility and link Cadwyn issue

Cadwyn is not compatible (yet) with FastAPI 0.128.0 - but the changes in
FastAPI are deliberately breaking so there is no way we are going to
support FastAPI >=0.128.0 unless there is a fix from Cadwyn (and we
also need to fix generating models for task-sdk to not use
private method that has been removed in FastAPI 0.128.0.

Therefore, it's better to upper-bind FastAPI rather than exclude it,
because there might be more versions of FastAPI released before
Cadwyn gets fixed.

Also the issue https://github.com/zmievsa/cadwyn/issues/316 has been
created, describing the root cause and suggested way how the issue
could be fixed in Cadwyn and comment on the reason for upper binding
was added to make sure we keep track of it.

* Apply suggestion from @potiuk